+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ## [Unreleased]
 ### Added
 - plugin.xml file to meet shopware code quality standards
+### Fixed
+- improved SW 5.6 support
+- read payment fee ordernumber and label from configuration
 
 ## [2.9.88] - 2019-09-17
 ### Added